Change ints to unsigned longs.

+++ src/object.c
@@ -30,7 +30,7 @@ ObjFunction *new_function(void) {
     return function;
 }
 
-ObjNative *new_native(NativeFn function, int arity) {
+ObjNative *new_native(NativeFn function, ulong arity) {
     ObjNative *native = ALLOCATE_OBJ(ObjNative, OBJ_NATIVE);
     native->function = function;
     native->arity = arity;
